reorganize ActiveUsers components
